+++ src/main/java/ru/aston/lobchevskaya_eyu/task1/interfaces/Discount.java
@@ -1,6 +1,8 @@
 package ru.aston.lobchevskaya_eyu.task1.interfaces;
 
+import java.math.BigDecimal;
+
 public interface Discount {
 
-    Long getDiscount();
+    BigDecimal getDiscount();
 }
